Corrected 3.4.x hotfix structure of ItemModifiedAppearance

+++ WowPacketParser/Store/Objects/Hotfixes/ItemModifiedAppearanceHotfix.cs
@@ -49,7 +49,7 @@ namespace WowPacketParser.Store.Objects
         public int? OrderIndex;
 
         [DBFieldName("TransmogSourceTypeEnum")]
-        public sbyte? TransmogSourceTypeEnum;
+        public int? TransmogSourceTypeEnum;
 
         [DBFieldName("VerifiedBuild")]
         public int? VerifiedBuild = ClientVersion.BuildInt;
